% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_categoricalDS.R
\name{check_categoricalDS}
\alias{check_categoricalDS}
\title{Value-set checks for categorical clinical variables}
\usage{
check_categoricalDS(df)
}
\arguments{
\item{df}{A character string giving the name of the server-side data
frame to check.}
}
\value{
Character vector of column names with values outside their
  permitted set.
}
\description{
Value-set checks for categorical clinical variables
}
