% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_dataDS.R
\name{check_missing_dataDS}
\alias{check_missing_dataDS}
\title{Compute percentage of missing values per column}
\usage{
check_missing_dataDS(df)
}
\arguments{
\item{df}{A character string giving the name of the server-side data
frame for which the percentage of missing values per column is to be
calculated.}
}
\value{
A named numeric vector where each element represents the percentage
of missing values in the corresponding column of \code{df}.
}
\description{
Computes the proportion of missing values (NA) for each column of a data frame.
The function returns a named numeric vector containing the percentage of
missing values for every variable.
}
