% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numericDS.R
\name{check_numericDS}
\alias{check_numericDS}
\title{Range/format checks for numeric clinical variables}
\usage{
check_numericDS(df)
}
\arguments{
\item{df}{A character string giving the name of the server-side data
frame to check.}
}
\value{
Character vector of column names failing their range/format check.
}
\description{
Range/format checks for numeric clinical variables
}
