% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2tAnalysisDS.R
\name{chisq_by_groupDS}
\alias{chisq_by_groupDS}
\title{Chi-square association test of several categorical variables with a group column}
\usage{
chisq_by_groupDS(df, vars, group_col, nfilter = 5)
}
\description{
Categorical analog of \code{kruskal_by_groupDS}, for
  Objective 2's between-cohort heterogeneity check. A variable is skipped
  (\code{NA}) if any contingency-table cell used is below \code{nfilter}.
}
