% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_categoricalDS.R
\name{.cdh_check_categorical_core}
\alias{.cdh_check_categorical_core}
\title{Value-set checks for categorical clinical variables (core logic)}
\usage{
.cdh_check_categorical_core(df)
}
\description{
Value-set checks for categorical clinical variables (core logic)
}
