% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numericDS.R
\name{.cdh_check_numeric_core}
\alias{.cdh_check_numeric_core}
\title{Range/format checks for numeric clinical variables (core logic)}
\usage{
.cdh_check_numeric_core(df)
}
\description{
Range/format checks for numeric clinical variables (core logic)
}
