% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_to_factorDS.R
\name{cast_to_factorDS}
\alias{cast_to_factorDS}
\title{Cast specified columns to factor}
\usage{
cast_to_factorDS(df, columns, newobj)
}
\arguments{
\item{df}{A data frame (already resolved).}

\item{columns}{A character vector (or "$"-joined string) of columns.}

\item{newobj}{Name under which the result is stored.}
}
\value{
\code{list(newobj=, columns_cast=)}.
}
\description{
Cast specified columns to factor
}
