% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_categoricalDS.R
\name{check_categoricalDS}
\alias{check_categoricalDS}
\title{Check validity of categorical variables}
\usage{
check_categoricalDS(df)
}
\arguments{
\item{df}{A data frame containing the categorical variables to be checked.}
}
\value{
A character vector with the names of columns containing invalid values.
  Returns an empty character vector if all checked variables are valid.
}
\description{
Verify that selected categorical variables in a data frame
  contain only predefined valid values.
}
